Well... this is... kinda big.
Fixed the round end / restart round or simply trigger the loading of the same map (via procon) crash.

Tweaked some probablities to try to enforce more vehicle based combat for the bots, so no more lonely chopper pilot without gunner, the air support looks very good, of course this is still configurable to keep a seat for a player, or we might aswell just adjust the probabilities (but for me these are the values I liked )

Commented out the annoying spawn selection in the node editor (it was useless for now anyway, but it was a pain if you selected a spawn by accident)

Fixed Operation Firestorm missing chopper for US in CQL

Fixed Sharqi for missing chopper for RU on Assault Large 64

Added vehicle usage for Nabandan Flats map CQL (still a WIP )

Added another class to discriminate between the attack helis / scouts and now the transport ones (venom, etc)

The 1000 value seems to be the sweetspot value for the Armed Gunship for me. (valued in Desert and Shield maps )

A lot of debugging and nil checks and other stuff.

For now hardcoded some limit for specific vehicles ( the Light AA for example), same way as for the artillery trucks.

With this current state the mod seems to be working without any crashes with many mods loaded at the same time:
My mod list:
fun-bots
VEManager
Bot-Civilianizer-Immersive
vu-compass
VU-Killstreak
vu-artillerystrike
vu-ks-attackheli
vu-ks-tank
no-flight-ceiling
Darkness-Unleashed
Vu-Loading-extended
VU-Noclip