Fixed the wrong bot states checking for vehicle actions when clearly they shouldn't have a vehicle related state since their player doesn't have any associated attachedControllable nor ControlledControllable entity.
Added proper state checking for the Vehicle States to transition as they should. Always validating first if the bot has died and then if it still has a proper controlledControllable or an attachedControllable.

Also enhanced a bit the call for ShouldExit for passengers, removed the AbortAttack call since that will be already properly handled within the _DoExitVehicle execution. Unless there was a need for triggering it before it gets executed. Open to discuss.

Added the FirstPassengerSeat property for missing vehicles such as the LittleBird helis, transport helis and No armor light vehicles such as the buggys

Please Hard test