Bot state fixes, Revert the VehicleMoveWhileShooting, it as ok as it was, my bad.
Removed the crossbow as default for the custom weapon sets for recon... it was really annoying to try to test stuff with the recon tending to use that a lot... each user particularly can re-enable them later.
Added missing FirstPassengerSeat definition for some vehicles that I think needed it, mainly vehicles with usable guns for their 2nd and 3rd seats.