This commit can be cherry picked to another branch if needed maybe.
The bots now have the random chance of calculating an offset on the fly while transiting the original path nodes, this helps to avoid them getting clumped and serves the purpose of giving them some more spacing and avoiding the feeling of them always following the same line like dumbs.

Added a lot of different checks and safety fallbacks for different scenarios ( them getting stuck being in the offset path, and so on )

Seems to be working flawlessly so far with the SideMovement disabled and the Jump While moving / shooting configs. ( I simply don't like it too much but its a personal decission )

# Conflicts:
#	ext/Server/Bot/BotMovement.lua